# openApiSchema.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Book API
paths:
  /books:
    get:
      summary: Get all books
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /books/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a book by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
    put:
      summary: Update a Book by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
    delete:
      summary: Delete a Book by ID
      responses:
        '204':
          description: No Content
        '404':
          description: Book not found
components:
  schemas:
    Book:
      type: object
      properties:
        book_id:
          type: integer
          readOnly: true
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        publication_year:
          type: integer
        isbn:
          type: string